Make tf_upgrade.py dependency free
Nothing else references the ast_edits, so it will make tf_upgrade.py much
easier to use if it's just absorbed. This change fixes #11217 where a whole
bunch of folks encountered difficulties for this very reason.
